Fix gnulib-cache.m4 removal code.

+++ autogen.sh
@@ -51,7 +51,9 @@ fi
 # Skip the gnulib-tool step if gnulib-tool was not found.
 if test -n "$GNULIB_TOOL"; then
   # In gettext-runtime:
-  mv -f gettext-runtime/gnulib-m4/gnulib-cache.m4 gettext-runtime/gnulib-m4/gnulib-cache.m4~
+  if test -f gettext-runtime/gnulib-m4/gnulib-cache.m4; then
+    mv -f gettext-runtime/gnulib-m4/gnulib-cache.m4 gettext-runtime/gnulib-m4/gnulib-cache.m4~
+  fi
   GNULIB_MODULES_RUNTIME_FOR_SRC='
   atexit
   basename
@@ -78,7 +80,9 @@ if test -n "$GNULIB_TOOL"; then
   $GNULIB_TOOL --dir=gettext-runtime --lib=libgrt --source-base=gnulib-lib --m4-base=gnulib-m4 --no-libtool --local-dir=gnulib-local --local-symlink \
     --import $GNULIB_MODULES_RUNTIME_FOR_SRC $GNULIB_MODULES_RUNTIME_OTHER
   # In gettext-tools:
-  mv -f gettext-tools/gnulib-m4/gnulib-cache.m4 gettext-tools/gnulib-m4/gnulib-cache.m4~
+  if test -f gettext-tools/gnulib-m4/gnulib-cache.m4; then
+    mv -f gettext-tools/gnulib-m4/gnulib-cache.m4 gettext-tools/gnulib-m4/gnulib-cache.m4~
+  fi
   GNULIB_MODULES_TOOLS_FOR_SRC='
   alloca-opt
   atexit
@@ -176,7 +180,9 @@ if test -n "$GNULIB_TOOL"; then
   $GNULIB_TOOL --dir=gettext-tools --lib=libgettextlib --source-base=gnulib-lib --m4-base=gnulib-m4 --makefile-name=Makefile.gnulib --libtool --local-dir=gnulib-local --local-symlink \
     --import $GNULIB_MODULES_TOOLS_FOR_SRC $GNULIB_MODULES_TOOLS_OTHER
   # In gettext-tools/libgettextpo:
-  mv -f gettext-tools/libgettextpo/gnulib-m4/gnulib-cache.m4 gettext-tools/libgettextpo/gnulib-m4/gnulib-cache.m4~
+  if test -f gettext-tools/libgettextpo/gnulib-m4/gnulib-cache.m4; then
+    mv -f gettext-tools/libgettextpo/gnulib-m4/gnulib-cache.m4 gettext-tools/libgettextpo/gnulib-m4/gnulib-cache.m4~
+  fi
   # This is a subset of the GNULIB_MODULES_FOR_SRC.
   GNULIB_MODULES_LIBGETTEXTPO='
   basename
